requirements.txt
flask==2.3.2
tensorflow==2.15.0
numpy==1.26.4
gdown==5.1.0
gunicorn==20.1.0
pillow==10.3.0